+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md)
 
 ## [Unreleased]
+### Added
+- Added new check-puppet-last-run-report.rb for new report file format that has replaced the last run summary file.
 
 ## [3.0.0] - 2020-03-24
 ### Breaking Change